Work to get an initial working simulator.
In these changes my assumption now is that we will support threading in general until we disable. I haven't implemented the code to handle when we don't have the threading available. We will need to add that logic, but should be as simple as just calling the code that would be called when the pop happens for the queue. We could handle this by  adding an abstraction layer that handles if we are using threading or not.